Backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
alembic
redis
rq
boto3
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pydantic-settings
python-dotenv
langgraph
langchain

